requirements.txt
altgraph==0.17.3
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2023.0
python-i18n==0.3.9
pywin32-ctypes==0.2.0
PyYAML==6.0
